+++ lib/src/constants.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/foundation.dart';
+
+const kTitleKey = const Key('FLUTTER_LOGIN_TITLE');
+const kRecoverPasswordIntroKey = const Key('RECOVER_PASSWORD_INTRO');
+const kRecoverPasswordDescriptionKey = const Key('RECOVER_PASSWORD_DESCRIPTION');
+const kDebugToolbarKey = const Key('DEBUG_TOOLBAR');