+++ lib/src/models/login_data.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/foundation.dart';
+import 'package:quiver/core.dart';
 
 class LoginData {
   final String name;
@@ -13,4 +14,13 @@ class LoginData {
   String toString() {
     return '$runtimeType($name, $password)';
   }
+
+  bool operator ==(Object other) {
+    if (other is LoginData) {
+      return name == other.name && password == other.password;
+    }
+    return false;
+  }
+
+  int get hashCode => hash2(name, password);
 }