+++ lib/src/providers/auth.dart
@@ -19,20 +19,12 @@ class Auth with ChangeNotifier {
     this.onSignup,
     this.onRecoverPassword,
     this.onConfirmRecover,
-    Auth previous,
-  }) {
-    if (previous != null) {
-      _mode = previous.mode;
-    }
-  }
-
-  Auth.empty()
-      : this(
-          onLogin: null,
-          onSignup: null,
-          onRecoverPassword: null,
-          previous: null,
-        );
+    email = '',
+    password = '',
+    confirmPassword = '',
+  })  : this._email = email,
+        this._password = password,
+        this._confirmPassword = confirmPassword;
 
   final AuthCallback onLogin;
   final AuthCallback onSignup;
@@ -62,5 +54,26 @@ class Auth with ChangeNotifier {
       mode = AuthMode.Login;
     }
     return mode;
+  }
+
+  String _email = '';
+  get email => _email;
+  set email(String email) {
+    _email = email;
+    notifyListeners();
+  }
+
+  String _password = '';
+  get password => _password;
+  set password(String password) {
+    _password = password;
+    notifyListeners();
+  }
+
+  String _confirmPassword = '';
+  get confirmPassword => _confirmPassword;
+  set confirmPassword(String confirmPassword) {
+    _confirmPassword = confirmPassword;
+    notifyListeners();
   }
 }