+++ lib/src/providers/login_messages.dart
@@ -28,7 +28,7 @@ class LoginMessages with ChangeNotifier {
   static const defaultLoginButton = 'LOGIN';
   static const defaultSignupButton = 'SIGNUP';
   static const defaultRecoverPasswordButton = 'RECOVER';
-  static const defaultRecoverPasswordIntro = 'Don\'t worry, happens to the best of us.';
+  static const defaultRecoverPasswordIntro = 'Reset your password here';
   static const defaultRecoverPasswordDescription =
       'We will send a password recovery code to your email.';
   static const defaultGoBackButton = 'BACK';